Bug 1260257 - New account Github only.
UI changes associated with disabling account creation for Persona.

This PR includes a change to CSS files and should not be deployed during peak traffic times because of bug 1259503.

I asked in bug 1265546 to have a page we can link to from the "?", right now it links to https://developer.mozilla.org/en-US/docs/MDN/Contribute/Howto/Create_an_MDN_account That page will also need to be updated but I'd like a more specific page.