bug 1277577 - Upgrade, move translate-toolkit
Update translate-toolkit from 1.13 to 2.0.0b2.  This upgrade adds Py3
compatibility, API changes, bug fixes, and feature updates.

We do not appear to use this package in production, so it is moved to
the local development requirements, and its requirements moved to
constraints.  The only documented usage is the podebug command, which is
given as an optional step in adding a new locale. This requirement is a
candidate for removal.